refactor to import with EngineSnapshot, xopp import is now asynchronous
